Make rawSwizzleInput!() faster
Now it's almost as fast as xgpuSwizzleRawInput!() which calls the C
implementation.  Not sure which one to keep long term.  Might be best to
call the C implementation if the requiste function/symbol exists in
the libxgpu shared library, or fall back to the Julia implementation if
it's not available.