Allow use with fewer antennas/channels than XGPU
This allows rawSwizzleInput! to swizzle input data for fewer antennas
and/or channels than XGPU was compiled for.  This is inefficient since
XGPU will be computing correlations that will not be valid/used, but it
allows for more flexibility in usage (e.g. it does not require a
rebuild of XGPU if some antennas are offline for a given observation).